pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test-ominimo"
version = "0.0.0"
description = "A small Python project created in the test_ominimo workspace"
authors = [ { name = "Emilio Barragán"} ]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3"
]

dependencies = [
    "dbt-core==1.10.13",
    "apache-airflow==3.1.0",
    "duckdb==1.4.1",
    "dbt-duckdb==1.9.6"
]

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
